+++ analytics/data/apiv1alpha/alpha_analytics_data_client.go
@@ -51,6 +51,7 @@ func defaultAlphaAnalyticsDataGRPCClientOptions() []option.ClientOption {
 		internaloption.WithDefaultMTLSEndpoint("analyticsdata.mtls.googleapis.com:443"),
 		internaloption.WithDefaultAudience("https://analyticsdata.googleapis.com/"),
 		internaloption.WithDefaultScopes(DefaultAuthScopes()...),
+		internaloption.EnableJwtWithScope(),
 		option.WithGRPCDialOption(grpc.WithDisableServiceConfig()),
 		option.WithGRPCDialOption(grpc.WithDefaultCallOptions(
 			grpc.MaxCallRecvMsgSize(math.MaxInt32))),